Distinct aries/w3c diddoc divergence
