+++ aries_vcx/src/handlers/connection/mediated_connection.rs
@@ -32,7 +32,7 @@ use messages::a2a::A2AMessage;
 use messages::protocols::basic_message::message::BasicMessage;
 use messages::protocols::connection::invite::Invitation;
 use messages::protocols::connection::request::Request;
-use messages::did_doc::DidDoc;
+use messages::did_doc::aries::diddoc::DidDoc;
 use messages::protocols::discovery::disclose::{Disclose, ProtocolDescriptor};
 
 #[derive(Clone, PartialEq)]
