+++ aries_vcx/src/protocols/connection/invitee/states/initial.rs
@@ -1,6 +1,6 @@
 use messages::protocols::connection::invite::Invitation;
 use messages::protocols::connection::problem_report::ProblemReport;
-use messages::did_doc::DidDoc;
+use messages::did_doc::aries::diddoc::DidDoc;
 use crate::protocols::connection::invitee::states::invited::InvitedState;
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
