+++ aries_vcx/src/protocols/connection/invitee/states/responded.rs
@@ -1,4 +1,4 @@
-use messages::did_doc::DidDoc;
+use messages::did_doc::aries::diddoc::DidDoc;
 use messages::protocols::connection::problem_report::ProblemReport;
 use messages::protocols::connection::request::Request;
 use messages::protocols::connection::response::Response;
