+++ aries_vcx/src/protocols/connection/inviter/states/complete.rs
@@ -1,6 +1,6 @@
 use std::clone::Clone;
 
-use messages::did_doc::DidDoc;
+use messages::did_doc::aries::diddoc::DidDoc;
 use messages::protocols::discovery::disclose::ProtocolDescriptor;
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
