+++ aries_vcx/tests/test_pool.rs
@@ -9,7 +9,7 @@ pub mod utils;
 #[cfg(feature = "pool_tests")]
 mod integration_tests {
     use aries_vcx::messages::protocols::connection::did::Did;
-    use aries_vcx::messages::did_doc::service_aries::AriesService;
+    use messages::did_doc::aries::service::AriesService;
     use aries_vcx::utils::constants::{DEFAULT_SCHEMA_ATTRS, SCHEMA_DATA};
     use aries_vcx::utils::devsetup::{SetupWalletPool, SetupProfile};
     use aries_vcx::common::keys::{get_verkey_from_ledger, rotate_verkey};
@@ -18,7 +18,7 @@ mod integration_tests {
     use std::sync::Arc;
     use std::thread;
     use std::time::Duration;
-    use messages::did_doc::service_aries_public::EndpointDidSov;
+    use messages::did_doc::service_didsov::EndpointDidSov;
 
     #[tokio::test]
     async fn test_open_close_pool() {
