+++ messages/src/did_doc/aries/mod.rs
@@ -1,0 +1,2 @@
+pub mod diddoc;
+pub mod service;
