+++ messages/src/did_doc/w3c/mod.rs
@@ -1,0 +1,3 @@
+pub mod service;
+pub mod diddoc;
+pub mod model;
