+++ messages/src/protocols/connection/response.rs
@@ -1,4 +1,4 @@
-use crate::did_doc::DidDoc;
+use crate::did_doc::aries::diddoc::DidDoc;
 use crate::a2a::message_family::MessageFamilies;
 use crate::a2a::message_type::MessageType;
 use crate::a2a::{A2AMessage, MessageId};
@@ -104,7 +104,7 @@ impl Default for ConnectionSignature {
 
 #[cfg(feature = "test_utils")]
 pub mod test_utils {
-    use crate::did_doc::test_utils::_did_doc_inlined_recipient_keys;
+    use crate::did_doc::aries::diddoc::test_utils::_did_doc_inlined_recipient_keys;
 
     use super::*;
 
@@ -166,7 +166,7 @@ pub mod test_utils {
 #[cfg(test)]
 #[cfg(feature = "general_test")]
 pub mod unit_tests {
-    use crate::did_doc::test_utils::*;
+    use crate::did_doc::aries::diddoc::test_utils::*;
     use crate::protocols::connection::response::test_utils::{_did, _response, _thread_id};
 
     use super::*;
